requirements.txt
about-time==4.2.1
alive-progress==3.1.4
certifi==2023.7.22
charset-normalizer==3.3.0
DateTime==5.2
grapheme==0.6.0
idna==3.4
progress==1.6
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
urllib3==2.0.7
zope.interface==6.1
